+++ src/marks/vscode_settings.py
@@ -3,6 +3,7 @@ from typing import Any
 from talon import Context, Module, actions
 from pathlib import Path
 from ..vendor.jstyleson import loads
+import traceback
 
 mod = Module()
 
@@ -78,3 +79,28 @@ class WindowsUserActions:
                 Path(f"{os.environ['APPDATA']}/VSCodium/User/settings.json"),
             ]
         )
+
+
+def vscode_get_setting_with_fallback(
+    key: str,
+    default_value: Any,
+    fallback_value: Any,
+    fallback_message: str,
+):
+    """Returns a vscode setting with a fallback in case there's an error
+
+    Args:
+        key (str): The key of the setting to look up
+        default_value (Any): The default value to return if the setting is not defined
+        fallback_value (Any): The value to return if there is an error looking up the setting
+        fallback_message (str): The message to show to the user if we end up having to use the fallback
+
+    Returns:
+        Any: The value of the setting or the default or fall back
+    """
+    try:
+        return actions.user.vscode_get_setting(key, default_value)
+    except Exception as e:
+        print(fallback_message)
+        traceback.print_exc()
+        return fallback_value
